% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscale.R
\name{autoscale}
\alias{autoscale}
\title{Mean-center and scale the column of a matrix}
\usage{
autoscale(x, xmi = NA, xmo = NA, robust = F)
}
\arguments{
\item{x}{matrix}

\item{xmi}{matrix indicating which values are missing in \code{x}}

\item{xmo}{value that represents missings in the returned matrix}

\item{robust}{logical indicating whether to use \code{IQR(column_x) / 1.35} or just the standard deviation (\code{sd}).}
}
\value{
a matrix in which the columns are mean-centered and scaled.#'
}
\description{
Mean-center and scale the column of a matrix
}
