% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmat.R
\name{bmat}
\alias{bmat}
\title{Create the so-called B matrix}
\usage{
bmat(odist, fdist)
}
\arguments{
\item{odist}{matrix of observed distances}

\item{fdist}{matrix of fitted distances}
}
\value{
the B matrix
}
\description{
An internal function int the \link[rCOSA]{smacof} algorithm.
}
