% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boxatt.R
\name{boxatt}
\alias{boxatt}
\title{Barplots of univariate attribute dispersion within groups}
\usage{
boxatt(data, imps, grps, main = "Consider a maintitle here",
  ylab = "Consider a label here", xlab = "Consider a label here",
  SCALE = TRUE, colors = palette(), ...)
}
\arguments{
\item{data}{data matrix obtained from COSA analysis}

\item{imps}{typically a vector with attribute indices ordered on their importance}

\item{grps}{...}

\item{main}{...}

\item{ylab}{...}

\item{xlab}{...}

\item{SCALE}{logical TRUE/FALSE, whether to scale the data}

\item{colors}{...}

\item{...}{named arguments to be passed to the default method.}
}
\value{
values of attribute importances in requested order.
}
\description{
This function plots the attribute weights over groups.
}
\examples{
make sure an at object is available...
}
\seealso{
\code{\link[COSA]{attimp}}
}

