% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxw.R
\name{boxw}
\alias{boxw}
\title{Barplots of attribute weights within groups}
\usage{
boxw(W, grpnr, attr, ...)
}
\arguments{
\item{W}{weights matrix obtained from COSA analysis}

\item{grpnr}{object row numbers of a specific group}

\item{attr}{attribute column numbers ordererd from important to less important. To find out the order use \code{attimp}.}

\item{...}{named arguments to be passed to the default method.}
}
\value{
values of attribute importances in requested order.
}
\description{
This function plots the attribute weights over groups.
}
\examples{
# make sure an at object is available...
}
\seealso{
\code{\\link[rCOSA]{attimp}}
}
