% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bplot.R
\name{bplot}
\alias{bplot}
\title{Barplot for the rCOSA results}
\usage{
bplot(
  x,
  xlab = "",
  ylab = "",
  zlim = c(0, 1),
  names = NULL,
  horiz = F,
  trans = 1,
  ntk = 20,
  names.size = 1
)
}
\arguments{
\item{x}{a vector or matrix of values}

\item{xlab}{label on the horizontal axis}

\item{ylab}{label on the vertical axis}

\item{zlim}{limits for the axis representing the height of the bars}

\item{names}{label on the vertical axis}

\item{horiz}{logical value that indicates whether the bars are shown horizontal or not (default is \code{FALSE}).}

\item{trans}{...}

\item{ntk}{number of equally spaced thicks on the barplot axis}

\item{names.size}{sizes of the names (expressed in values for \code{cex.names})}
}
\description{
Create a Barplot for rCOSA results
}
