% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmds.R
\name{cmds}
\alias{cmds}
\title{Classical metric multidimensional scaling}
\usage{
cmds(D, groupnr = NULL, main = NULL, colv = palette()[c(8, 2, 4, 3, 5, 6,
  7, 1)], ..., PLOT = TRUE)
}
\arguments{
\item{D}{dissimilarities/distances of class dist.}

\item{groupnr}{groupnr when known, each object can be given a number to which group it belongs}

\item{main}{title of the plot, by default there is no title.}

\item{colv}{vector containing color names for each group number}

\item{...}{arguments which can be parsed to either the \code{\link[mva]{hclust}} and \code{\link[stats]{cmdscale}} procedures.}

\item{PLOT}{whether to plot or not to plot, by default TRUE.}

\item{VERBOSE}{whether to give a messages on the criterion for each iteration, by default set to equal TRUE.}

\item{PLOT}{whether to plot or not to plot, by default TRUE.}
}
\value{
similar output as \code{\link[stats]{cmdscale}}
}
\description{
Classical metric multidimensional scaling which is the same as the cmdscale in the stats package. Which follows the analyis of Mardia (1978): Mardia, K.V. (1978) Some properties of classical multidimensional scaling. Communications on Statistics – Theory and Methods, A7, 1233–41.
}

