% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimdata.R
\name{dimdata}
\alias{dimdata}
\title{Retrieve the number of observations from a distance object}
\usage{
dimdata(data)
}
\arguments{
\item{data}{an object of class \code{dist}, or a vector with distances}
}
\value{
the number of observation in the data set
}
\description{
Retrieve the number of observations from a distance object
}
