% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Compute fit between two distance matrices}
\usage{
fit(delta, dd)
}
\arguments{
\item{delta}{observed distance matrix}

\item{dd}{euclidean distance matrix of a configuration}
}
\value{
the value for the fit (STRESS)
}
\description{
Compute fit between two distance matrices
}
