% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getclust.R
\name{getclust}
\alias{getclust}
\title{Identify Clusters in a Dendrogram}
\usage{
getclust(hc, ngr = 20, rec.col = "blue", old.col = "blue")
}
\arguments{
\item{hc}{an object of the type produced by \code{\link[COSA]{hierclust}} or \code{\link[stat]{hclust}}}

\item{ngr}{maximum number of clusters that can be selected}

\item{rec.col}{color of the most recently selected cluster}

\item{old.col}{color of the previously selected clusters.}
}
\value{
(Invisibly) returns a list where each element contains a vector of object identities contained in the respectively selected clusters.
}
\description{
Reads the position of the graphics pointer when the left mouse button is pressed. It then cuts the tree at the vertical position of the pointer and draws a colored rectangle (rec.col) around the cluster containing the horizontal position of the pointer, and identifies the objects in the corresponding cluster. This can be repeated up to a maximum of N times. Each press of the left mouse button selects a (different) cluster. The new cluster is colored rec.col and the color of the previous one is changed to old.col. When finished, pressing the right mouse button and selecting 'stop' returns control to the command window.
}
\note{
Make sure you have opened a plot device of your dendrogram already! The vector of object identities for each selected cluster are contained in \code{group[[1]]}, \code{group[[2]]}, \code{...}, \code{group[[NrG]]} in order of selection, where \code{NrG} is the number of selected groups (clicks of the left mouse button). Press Esc or click the right mouse button to end the selection procedure.
}
\examples{
# need to have an object hc produced by hierclust() or hclust()
# groups <- getclust(hc, NrG= 3)
}
\seealso{
\code{\link[COSA]{hierclust}}, \code{\link[stats]{hclust}}, and \code{\link[cluster]{pam}}
}

