% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hierclust}
\alias{hierclust}
\title{Hierarchical Clustering}
\usage{
hierclust(D, main = "", leaflab = c("perpendicular", "textlike", "none")[3],
  cex.main = NULL, denplot = TRUE, sub = "",
  xlab = "COSA dissimilarity clustering", ylab = "", method = "average",
  groupnr = NULL, colv = palette()[c(8, 2, 4, 3, 5, 6, 7, 1)], ...)
}
\arguments{
\item{D}{dissimilarities/distances of class dist.}

\item{main}{plot title}

\item{leaflab}{labels of the leaves, default is set to 'none'}

\item{cex.main}{redundant argument}

\item{denplot}{logical, whether to plot dendrogram. Default is set to \code{TRUE}.}

\item{sub}{subtitle}

\item{xlab}{label horizontal axis}

\item{ylab}{label vertical axis}

\item{groupnr}{when known, each object can be given a number to which group it belongs}

\item{colv}{vector containing color names for each group number}

\item{...}{arguments which can be parsed to either the \code{\link[mva]{hclust}} and \code{\link[stats]{dendrogram}} procedures.}

\item{the}{agglomeration method to be used. This should be (an unambiguous abbreviation of) one of \code{"ward"}, \code{"single"}, \code{"complete"}, \code{"average"}, \code{"mcquitty"}, \code{"median"} or \code{"centroid"}.}
}
\value{
Returns an invisible object of class dendrogram. If denplot is set to \code{TRUE} then the dendrogram is also plotted, and  when the groups are known one can ask for a colored dendrogram.
}
\description{
Performs a hierarchical cluster analysis on a set of dissimilarities using the R procedure \code{\link[stats]{hclust}} and \code{\link[stats]{dendrogram}}.
}

