% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowmat.R
\name{lowmat}
\alias{lowmat}
\title{Return matrix with only the lower diagonal}
\usage{
lowmat(udata)
}
\arguments{
\item{udata}{a distance object, or vector with distances}
}
\value{
matrix with only the lower diagonal
}
\description{
Return matrix with only the lower diagonal
}
