% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normdist.R
\name{normdist}
\alias{normdist}
\title{Normalize the distance}
\usage{
normdist(D)
}
\arguments{
\item{D}{A matrix of distances of class dist, or a vector or matrix containing distances.}
}
\value{
Returns a normalised version of D of similar class and mode.
}
\description{
Normalize the dissimilarities sucht that the squared norm of the Dissimilarities equals N, the number of objects.
}
