% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotimp.R
\name{plotimp}
\alias{plotimp}
\title{Replot attribute relevances with different plot parameters (after calling "attimp")}
\usage{
plotimp(
  attimp,
  atts = attimp$att[1:min(40, length(attimp$att))],
  imporder = F,
  donames = T,
  trans = 1,
  horiz = F,
  main = "Cluster",
  att.names = 1:length(attimp$att),
  names.size = 1,
  plot = T,
  ylim = c(0, max(imps)),
  ntick = 20
)
}
\arguments{
\item{attimp}{output from "attimp"}

\item{atts}{vector of attribute identities. This vector can be numeric, referencing the respective attribute numbers, or character, referencing the respective attribute names. If the latter, then the parameter "att.names" (see below) must be supplied.}

\item{imporder}{\code{TRUE} / \code{FALSE}  => do/don't plot relevance values in order of descending value.}

\item{donames}{\code{TRUE} / \code{FALSE} => do/don't show names of attributes on plot.}

\item{trans}{power used to transform importance scale.}

\item{horiz}{\code{TRUE} / \code{FALSE} => do/don't plot importances horizontally}

\item{main}{plot title}

\item{att.names}{vector of names for each attribute (ignored if donames= \code{FALSE}).}

\item{names.size}{expansion factor for names.}

\item{plot}{\code{TRUE} / \code{FALSE}  => do/don't display barplot.}

\item{ylim}{vertical range of plot.}

\item{ntick}{number of tick marks on importance axis.}
}
\value{
values of attribute importances in requested order.
}
\description{
This function allows replotting of the attribute revelances (importances) with different display parameters after calling "attimp", without having to recompute the relevance values. It also allows flexibility in how the results are displayed.
}
\examples{
# make sure an at object is available...
# plotimp(at)
# plotimp(at,1:100,trans=0.5,horiz=TRUE)
# plotimp(at,c(27,43,10,16))
# plotimp(at,c('age','income','occupation'),att.names=demographics)
}
\seealso{
\code{\link[rCOSA]{attimp}}
}
