% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCOSA-package.R
\docType{package}
\name{rCOSA-package}
\alias{rCOSA}
\alias{rCOSA-package}
\title{rCOSA: Clustering Objects on Subsets of Attributes (COSA)}
\description{
WARNING BETA VERSION. Implements the method of Friedman and
    Meulman (2004): "Clustering objects on subsets of attributes"; published in
    JRRS-B volume 66, part 4, pp. 815-849. Familiarity with this JRSS-B paper is
    assumed for those using this package.
}
\keyword{internal}
