% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotat.R
\name{rotat}
\alias{rotat}
\title{Rotate distance matrix (internal function rCOSA)}
\usage{
rotat(xx)
}
\arguments{
\item{xx}{distance matrix}
}
\value{
Rotated matrix in which the first column has the largest eigen value etc.
}
\description{
Rotate distance matrix (internal function rCOSA)
}
