% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/attimp.R
\name{attimp}
\alias{attimp}
\title{Compute and plot relevance of the attributes for selected clusters}
\usage{
attimp(X, group, range = 1:min(20, ncol(X)), times = 0, lX = if (class(X)
  == "data.frame") {     lX <- NULL } else {     lX <- 1 }, donames = FALSE,
  fast = FALSE, trans = 1, maximp = 20, xmiss = NULL, horiz = FALSE,
  main = "Cluster", att.names = 1:ncol(X), ord.names = FALSE,
  names.size = 1, ylim = c(0, wtave[o[1]]), ntick = 20, rlines = TRUE,
  targ = NULL, lwd = 1, ..., cache = FALSE)
}
\arguments{
\item{X}{data.frame or an input data matrix object of class matrix and mode numeric. Rows are objects and columns are attributes (variables).}

\item{range}{range of ordered importance values plotted (all are computed)}

\item{times}{number of randomly sampled groups for which ordered importances are also plotted}

\item{lX}{either an integer or a vector with as much elements as columns in \code{X}. This argument is ignored if \code{X} is a data.frame \cr}

\item{donames}{\code{TRUE} / \code{FALSE} => do/don't show names of attributes on plot.}

\item{fast}{\code{TRUE} / \code{FALSE} => do/don't use fast approximate calculation (less stable estimate). Should only be done with very large groups (when regular calculation takes way too long).}

\item{trans}{power used to transform important scale.}

\item{maximp}{maximum obtainable importance value}

\item{xmiss}{missing value flag (see the help file of \code{\link[COSA]{cosa}}).}

\item{horiz}{\code{TRUE}, \code{FALSE} => do/don't plot importances horizontally (ignored if time \code{> 0}).}

\item{main}{plot title}

\item{att.names}{vector of names for each attribute (ignored if donames = \code{FALSE}).}

\item{names.size}{expansion factor for names.}

\item{ylim}{vertical range of plot.}

\item{ntick}{number of tick marks on importance axis.}

\item{vector}{of identities of objects in the group}

\item{lX}{vector of attribute flags (see the help file of \code{\link[COSA]{cosa}}).
`}
}
\value{
\item{$imp}{vector of all attribute importances ordered in descending value.}
   \item{$att}{vector of attribute identities in descending order of importance.}
}
\description{
This function computes and displays the revelance (importance) of each attribute to the clustering of a selected group of objects. The relevance values are displayed in order of descending value. Optionally, on the same plot, are displayed corresponding ordered importances of randomly selected groups of objects of the same size.
}
\note{
the \code{lX} vector and/or the value of \code{xmiss} can be specified as attributes to the input data matrix before invoking attimp (see \code{\link[COSA]{cosa}})
}
\examples{
# make sure X and group are defined... For that, see getclust.
# at <- attimp(X, group[[2]])
# at <- attimp(X, group[[1]], 1:100, trans = 0.5, att.names = names, horiz = TRUE, main= 'Group 1')
# at <- attimp(X, group[[3]], 1:ncol(X), 10, main =' Group 3')
}
\seealso{
\code{\link[COSA]{dendro}}, \code{\link[stats]{hclust}}, \code{\link[COSA]{getclust}}, and \code{\link[COSA]{attvalues}}.
}

