% Generated by roxygen2 (4.0.2): do not edit by hand
\name{attvalues}
\alias{attvalues}
\title{Plot distributions of selected attributes within a specified cluster}
\usage{
attvalues(X, group, atts, lX = NULL, att.names = 1:ncol(X),
  group.name = "Group", xmiss = NULL, nbins = 25, all = F)
}
\arguments{
\item{X}{input data. Either a data.frame or a matrix in mode numeric. Rows are objects and columns are attributes (variables).}

\item{group}{vector of identities of objects in the group.}

\item{atts}{vector of identities of selected attributes.}

\item{lX}{vector of attribute flags (see \code{\link[COSA]{cosa}}).}

\item{att.names}{vector of names for each attribute.}

\item{group.name}{name for selected group (cluster).}

\item{xmiss}{missing value flag (see \code{\link[COSA]{cosa}}).}

\item{nbins}{(approXimate) number of histogram bins.}

\item{all}{\code{TRUE}, \code{FALSE} => do/don't show histograms for all data (third row).}
}
\value{
... STILL NEEDS A DESCRIPTION ...
}
\description{
This procedure plots a matrix of histograms. Each column pertains to one of the selected attributes. The first row shows the distribution for those objects within the specified group. The second row plots the corresponding distribution for those objects not in the specified group. Optionally (\code{all= TRUE}), a third row of histograms is produced showing the distribution of the corresponding attribute over all of the data. In all histograms, only nonmissing values of the respective attributes are plotted.
}
\note{
The \code{lX} vector and/or the value of \code{xmiss} can be specified as attributes to the input data matrix before invoking attimp (see \code{\link[COSA]{cosa}}).
}
\examples{
#acquire object at from the attimp function?
  #attvalues(X,gro	up[[2]],at$att[1:5])
  #attvalues(X,group[[2]],at$att[6:10],all=T)
  #attvalues(X,group[[4]],44,att.names='Att 44', group.name='Group 4')
}
\seealso{
\code{\link[COSA]{dendro}},\code{\link[stats]{hclust}}, \code{\link[COSA]{getclust}}, and \code{\link[COSA]{attvalues}}.
}

