% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxatt.R
\name{boxatt}
\alias{boxatt}
\title{Barplots of univariate attribute dispersion within groups}
\usage{
boxatt(
  data,
  imps,
  grps,
  main = "Consider a maintitle here",
  ylab = "Consider a label here",
  xlab = "Consider a label here",
  SCALE = TRUE,
  colors = palette(),
  ...
)
}
\arguments{
\item{data}{data matrix with numeric attributes}

\item{imps}{typically a vector with attribute indices ordered on their importance}

\item{grps}{list for which each component contains the index numbers (row numbers in the data) of each group in an integer vector.}

\item{main}{main title of the plot}

\item{ylab}{label vertical axis}

\item{xlab}{label horizontal axis}

\item{SCALE}{logical TRUE/FALSE, whether to scale the data}

\item{colors}{by default set to \code{palette()}}

\item{...}{named arguments to be passed to the default method.}
}
\value{
values of attribute importances in requested order.
}
\description{
This function plots the attribute weights over groups.
}
\examples{
# boxatt(data = ApoE3, imps = attWild$att[1:100],
#  #     grps =  list(1:18, 19:38),
#  #     main = "Wildtype (red) vs Transgenic (blue)",
#  #     ylab = 'Scaled attribute values',
#  #     xlab = '100 most important attributes',
#  #     colors = c('red','white','blue'))
}
\seealso{
\code{\link[rCOSA]{attimp}}
}
