% Generated by roxygen2 (4.0.2): do not edit by hand
\name{boxatt}
\alias{boxatt}
\title{Barplots of univariate attribute dispersion within groups}
\usage{
boxatt(X, grpnr, attr, ...)
}
\arguments{
\item{grpnr}{object row numbers of a specific group}

\item{attr}{attribute column numbers ordererd from important to less important. To find out the order use \code{attimp}.}

\item{...}{named arguments to be passed to the default method.}

\item{W}{weights matrix obtained from COSA analysis}
}
\value{
values of attribute importances in requested order.
}
\description{
This function plots the attribute weights over groups.
}
\examples{
#make sure an at object is available...
}
\seealso{
\code{\link[COSA]{attimp}}
}

