% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bplot.R
\name{bplot}
\alias{bplot}
\title{Title for this function}
\usage{
bplot(x, xlab = "", ylab = "", zlim = c(0, 1), names = NULL,
  horiz = F, trans = 1, ntk = 20, names.size = 1)
}
\arguments{
\item{x}{}

\item{xlab...}{}
}
\description{
Description of the function here
}

