% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmds.R
\name{cmds}
\alias{cmds}
\title{Classical (Metric) Multidimensional Scaling}
\usage{
cmds(
  D,
  groupnr = NULL,
  main = NULL,
  colv = palette()[c(8, 2, 4, 3, 5, 6, 7, 1)],
  PLOT = TRUE,
  ...
)
}
\arguments{
\item{D}{dissimilarities/distances of class dist.}

\item{groupnr}{groupnr when known, each object can be given a number to which group it belongs}

\item{main}{title of the plot, by default there is no title.}

\item{colv}{vector containing color names for each group number}

\item{PLOT}{whether to plot or not to plot, by default TRUE.}

\item{...}{arguments that can be parsed to \code{\link[stats]{cmdscale}} procedures.}
}
\value{
similar output as \code{\link[stats]{cmdscale}}
}
\description{
Classical metric multidimensional scaling which is the same as the cmdscale in the stats package. Which follows the analyis of Mardia (1978)
}
\references{
Mardia, K.V. (1978).  Some properties of classical multidimensional scaling.  Communications on Statistics - Theory and Methods, A7, 1233-41. \cr doi: \url{https://doi.org/10.1080/03610927808827707}
}
