% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cosaversion.R
\name{cosaversion}
\alias{cosaversion}
\title{Prints date and version number of COSA installation}
\usage{
cosaversion()
}
\value{
Returns the date of the last update of the COSA installion and the authors involved in creating this package
}
\description{
Prints date and version number of COSA installation
}

